chore: move controlPanel into time_table
