+++ superset-frontend/src/explore/components/controls/VizTypeControl/VizTypeControl.test.tsx
@@ -33,7 +33,7 @@ import {
   EchartsTimeseriesChartPlugin,
 } from '@superset-ui/plugin-chart-echarts';
 import { LineChartPlugin } from '@superset-ui/preset-chart-xy';
-import TimeTableChartPlugin from '../../../../visualizations/TimeTable/TimeTableChartPlugin';
+import TimeTableChartPlugin from '../../../../visualizations/TimeTable';
 import VizTypeControl, { VIZ_TYPE_CONTROL_TEST_ID } from './index';
 
 jest.useFakeTimers();
